增加指标版本控制

+++ coolGuard-access/src/main/java/cn/wnhyang/coolGuard/service/impl/IndicatorServiceImpl.java
@@ -10,20 +10,25 @@ import cn.wnhyang.coolGuard.constant.SceneType;
 import cn.wnhyang.coolGuard.context.AccessRequest;
 import cn.wnhyang.coolGuard.context.IndicatorContext;
 import cn.wnhyang.coolGuard.convert.IndicatorConvert;
+import cn.wnhyang.coolGuard.convert.IndicatorVersionConvert;
 import cn.wnhyang.coolGuard.entity.Chain;
 import cn.wnhyang.coolGuard.entity.Indicator;
+import cn.wnhyang.coolGuard.entity.IndicatorVersion;
 import cn.wnhyang.coolGuard.entity.PolicySet;
 import cn.wnhyang.coolGuard.enums.WinSize;
 import cn.wnhyang.coolGuard.indicator.AbstractIndicator;
 import cn.wnhyang.coolGuard.mapper.ChainMapper;
 import cn.wnhyang.coolGuard.mapper.IndicatorMapper;
+import cn.wnhyang.coolGuard.mapper.IndicatorVersionMapper;
 import cn.wnhyang.coolGuard.mapper.PolicySetMapper;
 import cn.wnhyang.coolGuard.pojo.PageResult;
 import cn.wnhyang.coolGuard.service.IndicatorService;
 import cn.wnhyang.coolGuard.util.IndicatorUtil;
+import cn.wnhyang.coolGuard.util.JsonUtils;
 import cn.wnhyang.coolGuard.util.LFUtil;
 import cn.wnhyang.coolGuard.vo.Cond;
 import cn.wnhyang.coolGuard.vo.IndicatorVO;
+import cn.wnhyang.coolGuard.vo.base.VersionSubmitVO;
 import cn.wnhyang.coolGuard.vo.create.IndicatorCreateVO;
 import cn.wnhyang.coolGuard.vo.page.IndicatorByPolicySetPageVO;
 import cn.wnhyang.coolGuard.vo.page.IndicatorPageVO;
@@ -40,6 +45,9 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+
+import static cn.wnhyang.coolGuard.exception.ErrorCodes.INDICATOR_IS_RUNNING;
+import static cn.wnhyang.coolGuard.exception.util.ServiceExceptionUtil.exception;
 
 /**
  * 指标表 服务实现类
@@ -59,11 +67,14 @@ public class IndicatorServiceImpl implements IndicatorService {
 
     private final ChainMapper chainMapper;
 
-    public IndicatorServiceImpl(List<AbstractIndicator> indicatorList, IndicatorMapper indicatorMapper, PolicySetMapper policySetMapper, ChainMapper chainMapper) {
+    private final IndicatorVersionMapper indicatorVersionMapper;
+
+    public IndicatorServiceImpl(List<AbstractIndicator> indicatorList, IndicatorMapper indicatorMapper, PolicySetMapper policySetMapper, ChainMapper chainMapper, IndicatorVersionMapper indicatorVersionMapper) {
         addIndicator(indicatorList);
         this.indicatorMapper = indicatorMapper;
         this.policySetMapper = policySetMapper;
         this.chainMapper = chainMapper;
+        this.indicatorVersionMapper = indicatorVersionMapper;
     }
 
     private void addIndicator(List<AbstractIndicator> indicatorList) {
@@ -74,17 +85,13 @@ public class IndicatorServiceImpl implements IndicatorService {
     @Transactional(rollbackFor = Exception.class)
     @CacheEvict(value = RedisKey.INDICATOR, allEntries = true)
     public Long createIndicator(IndicatorCreateVO createVO) {
+        // TODO 校验指标名
         Indicator indicator = IndicatorConvert.INSTANCE.convert(createVO);
         indicator.setCode(IdUtil.fastSimpleUUID());
         indicator.setReturnType(IndicatorUtil.getReturnType(indicator.getType(), indicator.getCalcField()));
-        indicator.setTimeSlice(WinSize.getWinSizeValue(createVO.getWinSize()));
+        indicator.setTimeSlice(WinSize.getWinSizeValue(indicator.getWinSize()));
+        indicator.setCondStr(JsonUtils.toJsonString(createVO.getCond()));
         indicatorMapper.insert(indicator);
-
-        String condEl = LFUtil.buildCondEl(createVO.getCond());
-        String iChain = StrUtil.format(LFUtil.INDICATOR_CHAIN, indicator.getCode());
-        chainMapper.insert(new Chain().setChainName(iChain).setElData(StrUtil.format(LFUtil.IF_EL, condEl,
-                LFUtil.INDICATOR_TRUE_COMMON_NODE,
-                LFUtil.INDICATOR_FALSE_COMMON_NODE)));
         return indicator.getId();
     }
 
@@ -92,16 +99,10 @@ public class IndicatorServiceImpl implements IndicatorService {
     @Transactional(rollbackFor = Exception.class)
     @CacheEvict(value = RedisKey.INDICATOR, allEntries = true)
     public void updateIndicator(IndicatorUpdateVO updateVO) {
+        // TODO hash确认是否真的有更改
         Indicator indicator = IndicatorConvert.INSTANCE.convert(updateVO);
         indicator.setReturnType(IndicatorUtil.getReturnType(indicator.getType(), indicator.getCalcField()));
         indicatorMapper.updateById(indicator);
-        String condEl = LFUtil.buildCondEl(updateVO.getCond());
-        String iChain = StrUtil.format(LFUtil.INDICATOR_CHAIN, indicator.getCode());
-        Chain chain = chainMapper.getByChainName(iChain);
-        chain.setElData(StrUtil.format(LFUtil.IF_EL, condEl,
-                LFUtil.INDICATOR_TRUE_COMMON_NODE,
-                LFUtil.INDICATOR_FALSE_COMMON_NODE));
-        chainMapper.updateById(chain);
     }
 
     @Override
@@ -109,15 +110,18 @@ public class IndicatorServiceImpl implements IndicatorService {
     @CacheEvict(value = RedisKey.INDICATOR, allEntries = true)
     public void deleteIndicator(Long id) {
         Indicator indicator = indicatorMapper.selectById(id);
+        IndicatorVersion indicatorVersion = indicatorVersionMapper.selectRunningByCode(indicator.getCode());
+        if (indicatorVersion != null) {
+            throw exception(INDICATOR_IS_RUNNING);
+        }
         indicatorMapper.deleteById(id);
-        chainMapper.deleteByChainName(StrUtil.format(LFUtil.INDICATOR_CHAIN, indicator.getCode()));
     }
 
     @Override
     public IndicatorVO getIndicator(Long id) {
         Indicator indicator = indicatorMapper.selectById(id);
         IndicatorVO indicatorVO = IndicatorConvert.INSTANCE.convert(indicator);
-        indicatorVO.setCond(getCond(indicator.getCode()));
+        indicatorVO.setCond(JsonUtils.parseObject(indicator.getCondStr(), Cond.class));
         return indicatorVO;
     }
 
@@ -127,7 +131,7 @@ public class IndicatorServiceImpl implements IndicatorService {
 
         PageResult<IndicatorVO> voPageResult = IndicatorConvert.INSTANCE.convert(indicatorPageResult);
 
-        voPageResult.getList().forEach(indicatorVO -> indicatorVO.setCond(getCond(indicatorVO.getCode())));
+        voPageResult.getList().forEach(indicatorVO -> indicatorVO.setCond(JsonUtils.parseObject(indicatorVO.getCondStr(), Cond.class)));
         return voPageResult;
     }
 
@@ -151,8 +155,45 @@ public class IndicatorServiceImpl implements IndicatorService {
     @Override
     public List<IndicatorVO> listIndicator() {
         List<IndicatorVO> indicatorVOList = IndicatorConvert.INSTANCE.convert(indicatorMapper.selectList());
-        indicatorVOList.forEach(indicatorVO -> indicatorVO.setCond(getCond(indicatorVO.getCode())));
+        indicatorVOList.forEach(indicatorVO -> indicatorVO.setCond(JsonUtils.parseObject(indicatorVO.getCondStr(), Cond.class)));
         return indicatorVOList;
+    }
+
+    @Override
+    @Transactional(rollbackFor = Exception.class)
+    public void commit(VersionSubmitVO submitVO) {
+        // 提交后就同时在version表中增加一条记录，表示该指标在运行状态
+        Indicator indicator = indicatorMapper.selectById(submitVO.getId());
+        // 1、更新当前指标为已提交
+        indicatorMapper.updateById(new Indicator().setId(submitVO.getId()).setStatus(Boolean.TRUE));
+        // 2、查询是否有已运行的，有版本+1，没有版本1
+        IndicatorVersion indicatorVersion = indicatorVersionMapper.selectRunningByCode(indicator.getCode());
+        int version = 1;
+        if (indicatorVersion != null) {
+            version = indicatorVersion.getVersion() + 1;
+            // 关闭已运行的
+            indicatorVersionMapper.updateById(new IndicatorVersion().setId(indicatorVersion.getId()).setStatus(Boolean.FALSE));
+        }
+        // 3、插入新纪录并加入chain
+        IndicatorVersion convert = IndicatorVersionConvert.INSTANCE.convert(indicator);
+        convert.setVersion(version);
+        convert.setVersionDesc(submitVO.getVersionDesc());
+        convert.setStatus(Boolean.TRUE);
+        indicatorVersionMapper.insert(convert);
+        // 4、更新chain
+        String iChain = StrUtil.format(LFUtil.INDICATOR_CHAIN, indicator.getCode());
+        String condEl = LFUtil.buildCondEl(convert.getCondStr());
+        if (chainMapper.selectByChainName(iChain)) {
+            Chain chain = chainMapper.getByChainName(iChain);
+            chain.setElData(StrUtil.format(LFUtil.IF_EL, condEl,
+                    LFUtil.INDICATOR_TRUE_COMMON_NODE,
+                    LFUtil.INDICATOR_FALSE_COMMON_NODE));
+            chainMapper.updateById(chain);
+        } else {
+            chainMapper.insert(new Chain().setChainName(iChain).setElData(StrUtil.format(LFUtil.IF_EL, condEl,
+                    LFUtil.INDICATOR_TRUE_COMMON_NODE,
+                    LFUtil.INDICATOR_FALSE_COMMON_NODE)));
+        }
     }
 
     @LiteflowMethod(value = LiteFlowMethodEnum.PROCESS_FOR, nodeId = LFUtil.INDICATOR_FOR_NODE, nodeType = NodeTypeEnum.FOR, nodeName = "指标for组件")
@@ -161,9 +202,9 @@ public class IndicatorServiceImpl implements IndicatorService {
         IndicatorContext indicatorContext = bindCmp.getContextBean(IndicatorContext.class);
         String appName = accessRequest.getStringData(FieldName.appName);
         String policySetCode = accessRequest.getStringData(FieldName.policySetCode);
-        List<Indicator> indicators = indicatorMapper.selectListByScenes(appName, policySetCode);
-        indicatorContext.setIndicatorList(ListUtil.toCopyOnWriteArrayList(IndicatorConvert.INSTANCE.convert(indicators)));
-        return indicators.size();
+        List<IndicatorVersion> indicatorVersionList = indicatorVersionMapper.selectRunningListByScenes(appName, policySetCode);
+        indicatorContext.setIndicatorList(ListUtil.toCopyOnWriteArrayList(IndicatorConvert.INSTANCE.convertVersion(indicatorVersionList)));
+        return indicatorVersionList.size();
     }
 
     @LiteflowMethod(value = LiteFlowMethodEnum.IS_ACCESS, nodeId = LFUtil.INDICATOR_COMMON_NODE, nodeType = NodeTypeEnum.COMMON)
